Implements under construction page
Adds an "under construction" page and redirects all non-allowed routes to it.

Introduces a conditional layout component to show or hide header and footer components based on the current route, excluding the "under construction" page and other predefined paths.

Updates middleware to redirect to the under construction page, and adds exceptions to the redirect to the under construction page, but redirects root to under construction as well.