Adds basic blog management API endpoints
Adds API endpoints for managing authors, categories, and posts.
This includes GET, POST, PUT, and DELETE methods for each entity,
with input sanitization and validation to ensure data integrity.

Also, configures Next.js to allow images from any hostname.

Creates a new React hook `useBlogPost` that fetches the blog post.
Creates `getCachedPosts` method that cache the blog posts.

Adds a new 'not-found' page that is displayed to the user when a blog post is not found.