Adds Impressum page
Adds a new page to display legal information and contact details for the organization.

This page includes details about the organization's name, legal form, registration information, address, contact information, and legal representative. It also contains a disclaimer regarding the content of external links and the information provided on the website.

Removes the old homepage.